+++ internal/features/periods/blocks/form_sections_templ.go
@@ -324,7 +324,7 @@ func FormSections(view dto.PeriodFormView, schedules []scheduledto.PersonWithSch
 				}()
 			}
 			ctx = templ.InitializeContext(ctx)
-			templ_7745c5c3_Err = sblocks.StudentMultiselect(url, "period", view.StudentOptions).Render(ctx, templ_7745c5c3_Buffer)
+			templ_7745c5c3_Err = sblocks.StudentMultiselectNew("period", view.StudentOptions, view.StudentIDs).Render(ctx, templ_7745c5c3_Buffer)
 			if templ_7745c5c3_Err != nil {
 				return templ_7745c5c3_Err
 			}